冗長なAPI実装を排除
