+++ src/spatial_id/flex_id/convert.rs
@@ -1,3 +1,4 @@
+use crate::SpatialId;
 use alloc::boxed::Box;
 
 use crate::{FlexId, RangeId, SingleId};
