+++ src/spatial_id/range_id/impls.rs
@@ -236,7 +236,9 @@ impl SpatialId for RangeId {
     }
 
     fn seconds_range(&self) -> (u64, u64) {
-        RangeId::seconds_range(self)
+        let unit = self.i.seconds();
+        let t = self.t();
+        (t[0] * unit, (t[1] + 1) * unit)
     }
 }
